增加了停车 倒车 发车函数

+++ B车（路况） - 副本/App/test.c
@@ -50,7 +50,7 @@ extern float speed_power;
 extern uint8 none_steerctrl; 
 extern int16 speedctrl_left,speedctrl_right; 
 int16 speedctrl_left_opp,speedctrl_right_opp; 
-uint8 last_stop = 0;  //终点停车标记位  0为不是停车
+uint16 last_stop = 0;  //终点停车标记位  0为不是停车
 extern float speed_forecast; //预测将要达到的速度（PWM）
 extern float speed_forecast_error; //预测将要达到的速度的偏差（差速）
